// libs/types/notification/notification.ts
import { NotificationGroup, NotificationStatus, NotificationType } from "../../enums/notification.enum";

export interface TotalCounter{
    total:number;
}
export interface Notification {
  _id: string;
  notificationType: NotificationType;
  notificationStatus: NotificationStatus;
  notificationGroup: NotificationGroup ;
  notificationTitle: string;
  notificationDesc?: string;
  authorId: string;
  receiverId: string;
  propertyId?: string;
  articleId?: string;
  createdAt: Date;
  updatedAt: Date;
}

export interface Notifications {
    list: Notification[];
    metaCounter: TotalCounter[];
}